Make crag interactive and extensive streamlining
This makes "crag" interactive if called without arguments.

This also moves all code from the script to the module, so it doesn't
need to be recompiled again.  The MAIN sub is now exported by the
module if :MAIN is specified.

This adds a dependency on Prompt, to allow for user interaction.  If
an editor with history support is found, then the history file will
be stored in ~/crag.history.

This adds a dependency on CodeUnit, which allows keeping any grammar
changes around between EVALs, and thus making the interactive experience
much better.

Added proper "plan"s to the test files.

This breaks a number of test in xt: apparently anything with "in" in
it, is now breaking when used non-interactively.